requirements.txt
anyio==3.6.1
autopep8==1.7.0
bcrypt==4.0.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
PyJWT==2.5.0
PyMySQL==1.0.2
python-dotenv==0.21.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
wincertstore==0.2
